Rename client.dart as trello_client.dart
