+++ lib/trello_sdk.dart
@@ -1,6 +1,5 @@
 export 'src/sdk/boards/boards.dart';
 export 'src/sdk/cards/cards.dart';
-export 'src/sdk/client.dart';
 export 'src/sdk/coordinates.dart';
 export 'src/sdk/errors.dart';
 export 'src/sdk/extensions/collapsable_either.dart';
@@ -12,5 +11,6 @@ export 'src/sdk/lists/lists.dart';
 export 'src/sdk/members/members.dart';
 export 'src/sdk/misc.dart';
 export 'src/sdk/trello_authentication.dart';
+export 'src/sdk/trello_client.dart';
 export 'src/sdk/trello_models.dart';
 export 'src/sdk/trello_object.dart';
